requirements.txt
agno>=1.3.3
fastapi>=0.115.12
openai>=1.75.0
psycopg2-binary>=2.9.10
pydantic>=2.11.3
requests>=2.32.3
uvicorn>=0.34.1
yfinance>=0.2.55
python-dotenv>=1.0.0
typing_extensions>=4.5.0 